refactor: identify CUDA-specific DLPack bridge
Make the private extension's scope explicit so future hardware adapters can provide distinct native bridges.
